Organize WebServer as go module before to add new functions
