+++ src/webserver/go.mod
@@ -1,0 +1,3 @@
+module historytracers
+
+go 1.18
